Adapt annotation patching
