+++ utbot-spring-analyzer/src/main/kotlin/org/utbot/spring/utils/SourceFinder.kt
@@ -2,10 +2,13 @@ package org.utbot.spring.utils
 
 import com.jetbrains.rd.util.getLogger
 import com.jetbrains.rd.util.info
+import org.springframework.context.annotation.ImportResource
+import org.utbot.common.patchAnnotation
 import org.utbot.spring.api.ApplicationData
 import org.utbot.spring.config.TestApplicationConfiguration
 import org.utbot.spring.configurators.ApplicationConfigurationType
 import java.io.File
+import java.nio.file.Path
 import kotlin.io.path.Path
 
 private val logger = getLogger<SourceFinder>()
@@ -18,11 +21,13 @@ class SourceFinder(
     fun findSources(): Array<Class<*>> = when (configurationType) {
         ApplicationConfigurationType.XmlConfiguration -> {
             logger.info { "Using xml Spring configuration" }
-            val configurationManager = ConfigurationManager(classLoader, TestApplicationConfiguration::class.java)
+
             // Put `applicationData.configurationFile` in `@ImportResource` of `TestApplicationConfiguration`
-            configurationManager.patchImportResourceAnnotation(Path(applicationData.configurationFile).fileName)
+            patchImportResourceAnnotation(Path(applicationData.configurationFile).fileName)
+
             arrayOf(TestApplicationConfiguration::class.java)
         }
+
         ApplicationConfigurationType.JavaConfiguration -> {
             logger.info { "Using java Spring configuration" }
             arrayOf(
@@ -37,4 +42,13 @@ class SourceFinder(
             "xml" -> ApplicationConfigurationType.XmlConfiguration
             else -> ApplicationConfigurationType.JavaConfiguration
         }
+
+    private fun patchImportResourceAnnotation(userXmlFilePath: Path) =
+        patchAnnotation(
+            classLoader,
+            TestApplicationConfiguration::class.java,
+            ImportResource::class,
+            "value",
+            arrayOf(String.format("classpath:%s", "$userXmlFilePath"))
+        )
 }
